move Header to src/layout directory
